moved this cell specific parsing stuff to a seperate class cause it's gotta alot of speciifcs to I assume conver alot of scenerios which I don't care about. a user should beable to implement there own interface for constructing the ICellData if they need something specific
